Update I1/5/6 pass logic in LVD/XLT
In all the inspection pass files in the LVD and XLT, change the logic so that it will not insert a bit into the pass bitshift unless the result location (where it's inserted) is after the trigger location.

Additionally, restructure (but don't change the logic) of the files so that all of the existing checks are split across multiple lines, to improve readability / follow the logic easier.